[#26] Refactor: 모델 타입 이름을 계층에 맞게 변경 (ex. Product -> ProductSummary)
